requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
pandas==2.1.4
transformers==4.35.0
torch==2.1.0
huggingface-hub==0.17.3
requests==2.31.0
accelerate==0.24.0 